+++ grpc/middleware/correlationid/opentracing/source_test.go
@@ -4,13 +4,11 @@ import (
 	"testing"
 
 	"context"
-	"net/http"
 
-	"github.com/goph/fxt/http/middleware/correlationid/opentracing"
+	"github.com/goph/fxt/grpc/middleware/correlationid/opentracing"
 	"github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/mocktracer"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestOpentracingSource_ExtractCorrelationID(t *testing.T) {
@@ -21,14 +19,9 @@ func TestOpentracingSource_ExtractCorrelationID(t *testing.T) {
 	ctx := context.Background()
 	ctx = opentracing.ContextWithSpan(ctx, span)
 
-	req, err := http.NewRequest("GET", "", nil)
-	require.NoError(t, err)
-
-	req = req.WithContext(ctx)
-
 	source := otcorrelationid.NewOpentracingSource()
 
-	correlationID := source.ExtractCorrelationID(req)
+	correlationID := source.ExtractCorrelationID(ctx)
 
 	assert.Equal(t, "cid", correlationID)
 }
@@ -41,14 +34,9 @@ func TestOpentracingSource_ExtractCorrelationID_RestrictedKey(t *testing.T) {
 	ctx := context.Background()
 	ctx = opentracing.ContextWithSpan(ctx, span)
 
-	req, err := http.NewRequest("GET", "", nil)
-	require.NoError(t, err)
-
-	req = req.WithContext(ctx)
-
 	source := otcorrelationid.NewOpentracingSource(otcorrelationid.RestrictedKey("correlationid"))
 
-	correlationID := source.ExtractCorrelationID(req)
+	correlationID := source.ExtractCorrelationID(ctx)
 
 	assert.Equal(t, "cid", correlationID)
 }