+++ grpc/middleware/correlationid/opentracing/store_test.go
@@ -4,13 +4,11 @@ import (
 	"testing"
 
 	"context"
-	"net/http"
 
-	"github.com/goph/fxt/http/middleware/correlationid/opentracing"
+	"github.com/goph/fxt/grpc/middleware/correlationid/opentracing"
 	"github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/mocktracer"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestOpentracingStore_StoreCorrelationID(t *testing.T) {
@@ -20,14 +18,9 @@ func TestOpentracingStore_StoreCorrelationID(t *testing.T) {
 	ctx := context.Background()
 	ctx = opentracing.ContextWithSpan(ctx, span)
 
-	req, err := http.NewRequest("GET", "", nil)
-	require.NoError(t, err)
-
-	req = req.WithContext(ctx)
-
 	store := otcorrelationid.NewOpentracingStore()
 
-	store.StoreCorrelationID(req, "cid")
+	store.StoreCorrelationID(ctx, "cid")
 
 	assert.Equal(t, "cid", span.BaggageItem("correlation_id"))
 }
@@ -39,14 +32,9 @@ func TestOpentracingStore_StoreCorrelationID_RestrictedKey(t *testing.T) {
 	ctx := context.Background()
 	ctx = opentracing.ContextWithSpan(ctx, span)
 
-	req, err := http.NewRequest("GET", "", nil)
-	require.NoError(t, err)
-
-	req = req.WithContext(ctx)
-
 	store := otcorrelationid.NewOpentracingStore(otcorrelationid.RestrictedKey("correlationid"))
 
-	store.StoreCorrelationID(req, "cid")
+	store.StoreCorrelationID(ctx, "cid")
 
 	assert.Equal(t, "cid", span.BaggageItem("correlationid"))
 }