+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - http: rewritten correlation ID middleware
+- grpc: rewritten correlation ID middleware
 
 
 ## [0.22.0] - 2018-04-26